+++ packages/svelte/tests/parser-legacy/samples/loose-invalid-block/input.svelte
@@ -1,0 +1,13 @@
+{#if }
+{:else if }
+{:else }
+{/if}
+
+{#each }
+{/each}
+
+{#snippet }
+{/snippet}
+
+{#snippet foo}
+{/snippet}